+++ extension/src/popup/components/sendPayment/SendTo/index.tsx
@@ -18,7 +18,7 @@ import { defaultAccountBalances } from "popup/views/Account";
 import {
   saveDestination,
   transactionDataSelector,
-} from "popup/ducks/transactionData";
+} from "popup/ducks/transactionSubmission";
 
 import { Input, Loader } from "@stellar/design-system";
